requirements/requirements.txt
pydantic
pyyaml
Fabric
paramiko-ng
globus_sdk
pandas
